Collapse census vectors

Collapse several census vectors into a single vector

collapse_census_vectors(data, vectors, aggregate = FALSE)

Arguments

data

Census data, from e.g. list_census_vectors or get_census_vectors_and_children

vectors

A data frame of existing census vectors (vector) and new "vectors" (new_vector) which they will be collapsed into. The vector field in data will be replaced with new_vector. If relevant, so will the label field, and the details field will have the label element replaced with the new vector element.

aggregate

Whether to aggregate (sum) existing values. Defaults to FALSE.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(cancensus)
#> Census data is currently stored temporarily.
#> 
#>  In order to speed up performance, reduce API quota usage, and reduce unnecessary network calls, please set up a persistent cache directory by setting the environment variable CM_CACHE_PATH= '<path to cancensus cache directory>' or  setting options(cancensus.cache_path = '<path to cancensus cache directory>')
#> 
#>  You may add this environment varianble to your .Renviron or add this option, together with your API key, to your .Rprofile.
#> 

couples_with_children <- tibble(vector = c("v_CA21_502", "v_CA21_505"), new_vector = "Couples with children")

list_census_vectors("CA21") %>%
  collapse_census_vectors(couples_with_children) %>%
  filter(vector %in% couples_with_children[["new_vector"]])
#> # A tibble: 2 × 7
#>   vector                type  label                units paren…¹ aggre…² details
#>   <chr>                 <fct> <chr>                <fct> <chr>   <chr>   <chr>  
#> 1 Couples with children Total Couples with childr… Numb… v_CA21… Additi… CA 202…
#> 2 Couples with children Total Couples with childr… Numb… v_CA21… Additi… CA 202…
#> # … with abbreviated variable names ¹​parent_vector, ²​aggregation